added recipe class, starting ui
Cleared up and reorganized code. However, didn't start. I will start 9/15/21 and continue through the week.
