+++ lib/my_main_page.dart
@@ -1,0 +1,30 @@
+import 'package:flutter/material.dart';
+
+class MyMainPage extends StatefulWidget {
+  MyMainPage({Key? key, required this.title}) : super(key: key);
+
+  final String title;
+
+  @override
+  _MyMainPageState createState() => _MyMainPageState();
+}
+
+class _MyMainPageState extends State<MyMainPage> {
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text(widget.title),
+      ),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            
+          ],
+        ),
+      ),
+    );
+  }
+}
