+++ lib/recipe.dart
@@ -1,0 +1,3 @@
+class Recipe{
+
+}
